in testing.
Fixed couple of bugs:Some objects are Null or Nulled, because of service context changed.

changed here the flow for powerupdates.
bug in this is involved in the binding and unbinding
also in checking for alarms is one active bug
partly commented out connection updates. -> too much toast messages
phone version of sd process embedded. Here raw data is downsampled to 250.

In dutch european law this program cannot be a primary medical app.